chore: migrate the /enterprise page to the app router
